pyproject.toml
[project]
name = "gym-collabsort"
version = "0.1.0"
description = "A Gymnasium environment for training agents on a collaborative sorting task"
readme = "README.md"
authors = [
    { name = "Baptiste Pesquet", email = "bpesquet@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gymnasium>=1.2.0",
    "pillow>=11.3.0",
    "pygame>=2.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

[tool.ruff.lint]
# Enhance checking. See https://docs.astral.sh/ruff/rules/#flake8-bugbear-b for details
extend-select = ["B"]

[tool.pytest.ini_options]
# Add code coverage reporting
addopts = "--cov=gym_collabsort --cov-report=term-missing"
